分拆相关代码
